+++ src/main/java/com/example/chatmessage/mapper/MessageMapper.java
@@ -8,10 +8,10 @@ import java.util.List;
 @Mapper
 public interface MessageMapper {
     @Select("select * from messages where chat_id = #{chatId}")
-    List<Message> list(Integer chatId);
+    List<Message> list(Long chatId);
 
     @Delete("delete from messages where message_id = #{messageId}")
-    void deleteMessage(Integer messageId);
+    void deleteMessage(Long messageId);
 
     @Insert("insert into messages (chat_id, user_id, message_content, audio, create_time, update_time) values (#{chatId}, #{userId}, #{messageContent}, #{audio}, #{createTime}, #{updateTime})")
     void addMessage(Message message);
